Main

Documentation

main :: IO ()

buildPuzzle

Arguments

:: String
Dimensions of the puzzle.

-> String
2D Array of all given spaces.

-> String
List of words.

-> Options
Given options from command line.

-> String
Output puzzles in readable format.

Given all the files, function will process the input using lines and will return all the puzzles in printable format.

All files are read only once. According to entry parameters, spaces which are 1x1, are (not) omitted.

\(O(n^2 \times m)^2\) where \(n\) is biggest number of words of given length and \(m\) is number of unique lengths

removeOnesDuplicities

Arguments

:: [[Space]]
All spaces present in puzzle.

-> [[Space]]
Output, duplicit 1x1 spaces are removed.

Will remove all duplicate 1x1 spaces.

Complexity: \(O(n + m^2)\), where \(n\) is the number of all different lengths of all spaces, \(m\) is the number of all 1x1 spaces. \(O(1)\), if 1x1 spaces are disabled.

removeDuplicities

Arguments

:: Eq a

=> [a]
Input list

-> [a]
Output list, multiple occurencies removed.

Will remove all occurences of given item from list. Selected item is prepended and therefore one instance of given item is kept.

Complexity: \(O(n^2)\), where n is the number of items in given list. For each item, new list without that item is created. Complexity of that is \(O(n)\).

removeAllOfItem

Arguments

:: Eq a

=> [a]
Input list.

-> a
Item to be removed.

-> [a]
List with all occurencies of given item removed.

Will remove all occurences of given item in given array.

Complexity: \(O(n)\), where n is number of elements in given list, as function will traverse whole list once, and will copy only items, which wont be deleted.

lineify

Arguments

:: String
Input string; first space will be replaced by "n"

-> String
Output

Replace all the spaces with "\n" character. Such string could then be used with lines function.

Complexity: \(O(n)\); whole string is traversed once.

loadWords

Arguments

:: String
Input string as read by hGetContents

-> Int
Maximum length of words

-> [[String]]
All words separated by their lengths accordingly into sublists.

Given string as-is from IO, function will use lines to parse these lines, sort given words and separate them accordingly by their length. Only words short enough are preserved, thus the filter is used.

\(O(n^3)\); \(n\) is the number of words divided by "\n"

sortSpaces

Arguments

:: [Space]
Input list of spaces.

-> Bool
Include 1x1 spaces?

-> [[Space]]
Spaces separated by their lengths into sublists.

Given the list of all spaces, and whether to ignore 1x1 spaces or not, function will sort and separate spaces of equal length.

\(O(n^2)\); \(n\) is the length of input array

importSpaces

Arguments

:: String
String of input spaces.

-> [Space]
List of all spaces found.

Given string of spaces, function use lines to parse the input and will call 'importSpaces\''.

importSpaces'

Arguments

:: [String]
Unparsed spaces

-> [String]
Backup; used by spacesFromColumns

-> Int
Which row is currently being parsed. Should be set by caller to 0.

-> [Space]
Parsed spaces.

Given the list of spaces in string representation, function will parse given lines and generate spaces accordingly.

\(O(n \times m)\); where \(n\) is number of rows and \(m\) is number of columns.

spacesFromColumns

Arguments

:: [String]
Unparsed spaces.

-> Int
Which column is currently being parsed. Should be set by caller to 0.

-> [Space]
Parsed spaces.

Given the list of spaces in string representation, function will parse given lines and generate spaces accordingly.

\(O(n \times m)\); where \(n\) is number of columns and \(m\) is number of rows.

printPuzzle :: [GraphNode] -> [[String]]

From list of GraphNode is taken puzzle 2D array and exported.

\(O(n)\); where \(n\) is number of nodes.

parseLine

Arguments

:: String
String to parse.

-> Int
Which row is selected.

-> Int
Start of the current space.

-> Int
Number of free "spaces" that current space has yet.

-> Char
Char to be ignored

-> [Space]

Function will parse given line, where n and input Lexeme are both ignored regarding wether there is a wall or not. Any other char will be recognized as non-wall char.

\(O(n)\); where \(n\) is length of input string.

parseColumn

Arguments

:: String
String to parse.

-> Int
Which row is selected.

-> Int
Start of the current space.

-> Int
Number of free "spaces" that current space has yet.

-> Char
Char to be ignored

-> [Space]

Function will parse given line, where n and input Lexeme are both ignored regarding wether there is a wall or not. Any other char will be recognized as non-wall char.

\(O(n)\); where \(n\) is length of input string.

startSolvingPuzzle

Arguments

:: Maybe PuzzleState
Initial puzzle state. If value equals Nothing, empty list is returned.

-> [[String]]
List of available words.

-> [[Space]]
List of available spaces.

-> [GraphNode]
List of puzzles, where all spaces are filled.

Given initial PuzzleState, function will begin solving the puzzle using generateLevel.

createPuzzleState

Arguments

:: Maybe Int
Number of rows

-> Maybe Int
Number of columns

-> Maybe PuzzleState
New PuzzleState

Will create empty PuzzleState with no Spaces remaining.

No. of Spaces can be modified with changeSpaces. If value in rows or columns was incorrect, will result in Nothing

\(O(r \times c)\); \(r\) is number of rows and \(c\) is number of columns

changeSpaces

Arguments

:: Maybe PuzzleState
Old PuzzleState

-> Int
New space count

-> Maybe PuzzleState
New PuzzleState

Will change no. of Spaces in given puzzle.

\(O(1)\)

insertCol

Arguments

:: [[Char]]
Input (2D array).

-> [Char]
Word to insert.

-> Int
Column, to which the word sohuld be inserted.

-> Int
Position in column, at which the word should be inserted.

-> [[Char]]
Output.

Function insertCol will get 2D array of chars and will write down provided word into selected column at given position.

\(O(n)\), where \(n\) is the length of given list

insertRow

Arguments

:: [[Char]]
Input (2D array).

-> [Char]
Word, which should be inserted.

-> Int
Row, to which the word should be inserted.

-> Int
Position, at which the word should be inserted.

-> [[Char]]
Output.

Function insertRow will get 2D array of chars and will write down provided word into selected row at given position.

\(O(n)\), where \(n\) is the length of given list

dummy :: [[Char]]

peekWithSpace :: PuzzleState -> Space -> String

Peek sublist from given PuzzleState with provided Space.

>>> peekWithSpace (Puzzle 0 ["1234", "5678", "9abc", "def0"]) (Space (0,0) (2,0))
"123"

\(O(n + m)\), where \(n\) is the number of rows is selected and \(m\) is the length of the word. If row is being selected.

\(O (n \times m)\) is !! operator and \(m\) is the length of the word. If column is being selected.

peekSubColumn

Arguments

:: [[Char]]
Input

-> Int
Count; length of word

-> Int
Which column

-> String
Output

Will peek subcolumn from given column, one piece after another.

\(O(n * m)\), where \(n\) is !! operator and \(m\) is the length of the word

insertForSpace

Arguments

:: PuzzleState
Input (2D array with counter of inserted words).

-> [Char]
Word, which should be inserted.

-> Space
Space where to insert.

-> PuzzleState
Output.

Given Space and PuzzleState, function will delegate inserting the word to either insertRow or insertCol.

replaceItem

Arguments

:: Ord a

=> [a]
Input array.

-> [a]
Replacement array.

-> Int
Where to replace.

-> [a]
Output.

Function will replace part of given array at given position.

\(O(n)\); where \(n\) is the length of the input list

data Space

For representation of space, which will be filled during puzzle solving.

List of spaces is maintained during entire execution of the program, where unfilled spaces are still kept in the list. Filled ones are removed from the list and word is filled into puzzle 2D array.

Constructors

Space (Int, Int) (Int, Int)

spaceLen :: Space -> Int

Function for getting length of space with given wordspace.

Positive values means rows, negative values means columns.

\(O(1)\), only constant calculations are used.

rollArray :: [a] -> [a]

Will roll array one step to the left. >>> rollArray [1,2,3] [2,3,1]

\(O(n)\) where \(n\) is length of given list.

qsort

Arguments

:: [a]
Input

-> (a -> Int)
Function used to filter out items during sorting. Quantity, on which the sorting is based.

-> [a]
Sorted array.

Funcion will sort given list of words/spaces by their length, descending. >>> sort ["aaa", "zzzz", "tt"] length ["zzzz","aaa","tt"]

\(O(n^2) \times f \times log(n))\), where \(n\) is the length of given array and \(f\) is the comlexity of filtering function

myFilter

Arguments

:: (Int -> Int -> Bool)
Function used for ordering items.

-> (a -> Int)
Qualitative function, which is used for sorting.

-> a
First input.

-> a
Second input.

-> Bool
Whether or not was item filtered out.

Will filter out all items accoring to provided functions.

\(O(f + c)\) where both \(f\) and \(c\) are complexities of provided functions.

listsSeparate :: [a] -> (a -> Int) -> [[a]]

Following method will create groups of wordsspaces with equal length, assuming, that the wordsspaces are already sorted. Leaving out empty lists where no words/spaces of given length were found. >>> listsSeparate ["hello", "world", "how", "are", "you", "?"] length [["hello","world"],[],["how","are","you"],[],["?"]]

\(O(n^2 \times f)\), where \(n\) is length of input list and \(f\) is complexity of given function.

listSeparate' :: [a] -> (a -> Int) -> Int -> [[a]]

Helper function for listsSeparate

\(O(n^2 \times f)\), where \(n\) is length of input list and \(f\) is complexity of given function.

data GraphNode

Structure for encapsulating PuzzleState together with list of remaining words and spaces. !!

Structure basically represents one node of graph, which is traversed during the process of solving puzzle.

Constructors

Node PuzzleState [[String]] [[Space]]

data PuzzleState

For encapsulating actual puzzle 2D array and number of remaining empty spaces.

Constructors

Puzzle Int [[Char]]

separateLength :: [a] -> (a -> Int) -> Int -> [a]

Will separate all items of given length.

>>> separateLength ["ahoj", "svet", "ako", "mas", "sa", "ty", "?", "!"] length 4
["ahoj","svet"]

\(O(n \times f)\), where \(n\) is length of input list and \(f\) is complexity of given function

emptyPuzzle :: PuzzleState

dummyNode :: GraphNode

generateLevel :: GraphNode -> Int -> [GraphNode]

For given GraphNode function will generate all possible and valid subsequent nodes.

Behaviour of this function altogether with extendLevel is similiar to DFS algorithm and therefore its complexity is:

\(O(|V| + |E|)\), where \(V\) is \(n^2 \times m\) where \(n\) is biggest number of words of given length and \(m\) is number of unique lengths and \(E\) is \(V - 1\)

extendLevel :: [GraphNode] -> [GraphNode]

For all given nodes, function will generate new level using generateLevel

\(O(n^2)\), where \(n\) is the number of nodes in list

canBeReplaced

Arguments

:: String
Template.

-> String
Replacement string.

-> Bool
Result.

Determines, wether the given string could be replaced with second string, with . symbol being a wildcard.

>>> canBeReplaced "a.." "abc"
True

>>> canBeReplaced "aa." "abc"
False

\(O(n^2)\) where \(n\) is the length of the second, replacement string.

canBeInsertedPuzzle :: PuzzleState -> Space -> String -> Bool

Will determine, wether given string could be inserted at given Space into given PuzzleState.

\(O(n + c)\), where \(n\) is the length of the string and \(c\) is complexity of peekWithSpace.